% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCluster.R
\name{fCluster}
\alias{fCluster}
\title{Returns a partialised clustering function that can be used to construct an importance tree in \code{partim}}
\usage{
fCluster(type = c("diana", "agnes", "pam"), ..., balanced = FALSE)
}
\arguments{
\item{type}{the function from the \code{cluster} package to use for clustering. Default: 'diana'.}

\item{...}{additional arguments passed to the cluster function.}

\item{balanced}{boolean. When TRUE a balanced tree is constructed. When FALSE (the default) tree splits are determined using the clustering algorithm.}
}
\value{
A vector of cluster assignments for each variable in \code{x}.
}
\description{
\code{fCluster} is a wrapper for functions in the \code{cluster} package (\code{agnes}, \code{diana}, \code{pam}) and can be passed to \code{partim} to construct different types of hierarchical partitioning trees.
}
\details{
\code{fCluster} creates a partialised function which only requires input data (\code{x}) and splits this data into 2 clusters using the preset clustering algorithm.

Clustering is performed using a distance representation of \code{x}, such that distance matrix \eqn{d} is given by

\eqn{d = ||1 - cor(x)^2||_2}

The default approach partitions this distance matrix into 2 clusters at each recursive split along the tree using the Divisive Analysis algorithm (\code{cluster::diana}).
}
\examples{
data <- MASS::Boston

# Calculate importance using balanced single-linkage tree
partim(medv ~ ., data, fCluster = fCluster("agnes", method = "single", balanced = TRUE))

}
\references{
Maechler, M., Rousseeuw, P., Struyf, A., Hubert, M., Hornik, K.(2022).  cluster: Cluster Analysis Basics and Extensions. R package version 2.1.4.
}
\seealso{
\code{\link{partim}}, \code{\link{fEstimate}}
}
\author{
Johann Pfitzinger
}
