% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fEstimate.R
\name{fEstimate}
\alias{fEstimate}
\title{Returns a partialised estimation function that can be used for both estimation and explanation functions in \code{partim}}
\usage{
fEstimate(type = c("lm", "glmnet", "nnet"), ...)
}
\arguments{
\item{type}{the function to use for model fitting. Default: 'lm'.}

\item{...}{additional arguments passed to the fitting function.}
}
\value{
An object with a 'fitted.values' attribute.
}
\description{
\code{fEstimate} is a wrapper for \code{stats::lm}, \code{glmnet::glmnet} and \code{nnet::nnet}, and can be passed to \code{partim} to be used for model estimation and/or explanations.
}
\details{
\code{fEstimate} creates a partialised function which only requires input data (\code{x} and \code{y}) and returns an object with a 'fitted.values' attribute.

When \code{type = 'lm'}, a linear regression model is fitted. Additional arguments such as weights can be passed to \code{...}
When \code{type = 'glmnet'} a regularized regression model (Elastic Net, Lasso, Ridge) can be fitted. Additional arguments such as 'lambda' or 'alpha' can be passed to \code{...}
When \code{type = 'nnet'} a simple neural network model is fitted. Additional arguments such as 'size' can be passed to \code{...}
}
\examples{
data <- MASS::Boston

# Calculate importance of a Lasso regression
fEst <- fEstimate("glmnet", alpha = 1, lambda = 1)
partim(medv ~ ., data, fEstimate = fEst, fExplain = fEstimate("lm"))

}
\references{
Friedman J, Tibshirani R, Hastie T (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.”
\emph{Journal of Statistical Software}, \emph{33}(1), 1-22. doi:10.18637/jss.v033.i01 \url{https://doi.org/10.18637/jss.v033.i01}.

Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics with S.
Fourth Edition. Springer, New York. ISBN 0-387-95457-0
}
\seealso{
\code{\link{partim}}, \code{\link{fCluster}}
}
\author{
Johann Pfitzinger
}
